remove lib/vec.rs
