+++ lib/flux-alloc/src/vec/mod.rs
@@ -1,14 +1,17 @@
 use std::{
     alloc::{Allocator, Global},
-    // ops::{Index, IndexMut},
-    // slice::{Iter, SliceIndex},
+    ops::{Index, IndexMut},
+    slice::SliceIndex,
 };
 
 use flux_attrs::*;
 
+//---------------------------------------------------------------------------------------
 #[extern_spec]
 #[refined_by(len: int)]
 struct Vec<T, A: Allocator = Global>;
+
+//---------------------------------------------------------------------------------------
 
 #[extern_spec]
 impl<T> Vec<T> {
@@ -27,3 +30,29 @@ impl<T, A: Allocator> Vec<T, A> {
     #[flux::sig(fn(self: &mut Vec<T, A>[@n]) -> Option<T> ensures self: Vec<T, A>[if n > 0 { n-1 } else { 0 }])]
     fn pop(&mut self) -> Option<T>;
 }
+
+//---------------------------------------------------------------------------------------
+
+#[extern_spec]
+impl<T, I: SliceIndex<[T]>, A: Allocator> Index<I> for Vec<T, A> {
+    #[flux::sig(fn(&Vec<T, A>[@len], {I[@idx] | <I as SliceIndex<[T]>>::in_bounds(idx, len)}) -> _)]
+    fn index(z: &Vec<T, A>, index: I) -> &<I as SliceIndex<[T]>>::Output;
+}
+
+#[extern_spec]
+impl<T, I: SliceIndex<[T]>, A: Allocator> IndexMut<I> for Vec<T, A> {
+    #[flux::sig(fn(&mut Vec<T,A>[@len], {I[@idx] | <I as SliceIndex<[T]>>::in_bounds(idx, len)}) -> _)]
+    fn index_mut(z: &mut Vec<T, A>, index: I) -> &mut <I as SliceIndex<[T]>>::Output;
+}
+
+//---------------------------------------------------------------------------------------
+
+#[extern_spec]
+impl<'a, T, A: Allocator> IntoIterator for &'a Vec<T, A> {
+    #[flux::sig(fn (&Vec<T, A>[@n]) -> <&Vec<T, A> as IntoIterator>::IntoIter[0,n])]
+    fn into_iter(v: &'a Vec<T, A>) -> <&'a Vec<T, A> as IntoIterator>::IntoIter;
+}
+
+#[extern_spec]
+#[flux::assoc(fn with_size(self: Self, n:int) -> bool { self.len == n })]
+impl<T> FromIterator<T> for Vec<T> {}
