+++ tests/tests/neg/surface/iter_vec00.rs
@@ -10,10 +10,8 @@ use std::{
 
 use flux_rs::extern_spec;
 
+extern crate flux_alloc;
 extern crate flux_core;
-
-#[path = "../../lib/vec.rs"]
-mod vec;
 
 #[flux::sig(fn(bool[true]))]
 pub fn assert(_b: bool) {}
