+++ tests/tests/neg/surface/vec02.rs
@@ -2,10 +2,8 @@
 
 use std::ops::Index;
 
+extern crate flux_alloc;
 extern crate flux_core;
-
-#[path = "../../lib/vec.rs"]
-mod vec;
 
 // ---------------------------------------------------------------------------------------
