+++ tests/tests/pos/surface/closure11.rs
@@ -1,10 +1,8 @@
 #![feature(step_trait, allocator_api)]
 #![allow(unused)]
 
+extern crate flux_alloc;
 extern crate flux_core;
-
-#[path = "../../lib/vec.rs"]
-mod vec;
 
 #[flux_rs::sig(fn (bool[true]))]
 pub fn assert(b: bool) {}
