+++ tests/tests/pos/surface/iter02.rs
@@ -3,10 +3,8 @@
 
 use std::{iter::Enumerate, slice::Iter};
 
+extern crate flux_alloc;
 extern crate flux_core;
-
-#[path = "../../lib/vec.rs"]
-mod vec;
 
 #[flux::sig(fn(bool[true]))]
 pub fn assert(_b: bool) {}
